+++ src/remote/activitypub/misc/get-note-html.ts
@@ -1,5 +1,5 @@
 import { INote } from '../../../models/note';
-import { toHtml } from '../../../mfm/toHtml';
+import { toHtml } from '../../../mfm/to-html';
 import { parse } from '../../../mfm/parse';
 
 export default function(note: INote) {
