+++ src/remote/activitypub/renderer/person.ts
@@ -2,7 +2,7 @@ import renderImage from './image';
 import renderKey from './key';
 import config from '../../../config';
 import { ILocalUser } from '../../../models/user';
-import { toHtml } from '../../../mfm/toHtml';
+import { toHtml } from '../../../mfm/to-html';
 import { parse } from '../../../mfm/parse';
 import DriveFile from '../../../models/drive-file';
 import { getEmojis } from './note';
